Exception handling

+++ src/main/java/com/devbraga/bragacommerce/services/ProductService.java
@@ -3,10 +3,16 @@ package com.devbraga.bragacommerce.services;
 import com.devbraga.bragacommerce.dto.ProductDTO;
 import com.devbraga.bragacommerce.entities.Product;
 import com.devbraga.bragacommerce.repository.ProductRepository;
+import com.devbraga.bragacommerce.services.exceptions.DatabaseException;
+import com.devbraga.bragacommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
+import org.hibernate.boot.model.relational.Database;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -17,7 +23,8 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id){
-       Product product = productRepository.findById(id).get();
+       Product product = productRepository.findById(id).orElseThrow(
+               () -> new ResourceNotFoundException("Recurso não encontrado"));
        return new ProductDTO(product);
     }
 
@@ -39,10 +46,15 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto){
-        Product entity = productRepository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = productRepository.save(entity);
-        return new ProductDTO(entity);
+        try {
+            Product entity = productRepository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = productRepository.save(entity);
+            return new ProductDTO(entity);
+        }
+        catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
     //Método auxiliar para copiar os atributos da minha entidade para sua classe DTO
@@ -53,8 +65,16 @@ public class ProductService {
         entity.setImgUrl(dto.getImgUrl());
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
-       productRepository.deleteById(id);
+        if(!productRepository.existsById(id)){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            productRepository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e){
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 }
